requirements.txt
asyncio
playwright==1.50.0
pandas==2.2.3
ddddocr==1.5.6
pillow==11.1.0
